+++ components/Confirm/style/index.tsx
@@ -1,4 +1,3 @@
 import '../../style/';
 import '../../Modal/style/';
-import '../../Button/style/';
 import './index.scss';