+++ components/Modal/style/index.tsx
@@ -1,3 +1,4 @@
 import '../../style/';
 import '../../Mask/style/';
+import '../../Icon/style/';
 import './index.scss';