+++ examples/pages/ButtonPage.jsx
@@ -16,32 +16,38 @@ class Page extends Component {
           <Panel>
             <Panel.Header title="基本" />
             <Panel.Body>
-              <Button theme="primary">普通按钮</Button>
-              <Button theme="primary" block>块级按钮</Button>
-              <Button theme="primary" block active>激活状态的按钮</Button>
-              <Button theme="primary" block disabled>禁用状态的按钮</Button>
+              <Button>default</Button>
+              <Button theme="primary">primary</Button>
+            </Panel.Body>
+          </Panel>
+
+          <Panel>
+            <Panel.Header title="块级按钮" />
+            <Panel.Body>
+              <Button block>default</Button>
+              <Button block disabled shape="radius">disabled default</Button>
+              <Button block theme="primary">primary</Button>
+              <Button block disabled theme="primary">disabled primary</Button>
             </Panel.Body>
           </Panel>
 
           <Panel>
             <Panel.Header title="幽灵按钮" />
             <Panel.Body>
-              <Button block bordered>幽灵按钮</Button>
-              <Button block bordered active>激活状态的按钮</Button>
-              <Button block bordered disabled>禁用状态的按钮</Button>
-              <Button theme="primary" block bordered>幽灵按钮</Button>
-              <Button theme="primary" block bordered active>激活状态的按钮</Button>
-              <Button theme="primary" block bordered disabled>禁用状态的按钮</Button>
+              <Button block bordered>default</Button>
+              <Button block bordered disabled>disabled default</Button>
+              <Button block bordered theme="primary">primary</Button>
+              <Button block bordered disabled theme="primary">disabled primary</Button>
             </Panel.Body>
           </Panel>
 
           <Panel>
             <Panel.Header title="主题" />
             <Panel.Body>
-              <Button>default</Button>
+              <Button theme="default">default</Button>
               <Button theme="primary">primary</Button>
-              <Button theme="info">info</Button>
               <Button theme="success">success</Button>
+              <Button theme="info">info</Button>
               <Button theme="warning">warning</Button>
               <Button theme="error">error</Button>
             </Panel.Body>
@@ -50,21 +56,19 @@ class Page extends Component {
           <Panel>
             <Panel.Header title="大小" />
             <Panel.Body>
-              <Button theme="primary" size="xl">xl</Button>
               <Button theme="primary" size="lg">lg</Button>
               <Button theme="primary">md</Button>
               <Button theme="primary" size="sm">sm</Button>
-              <Button theme="primary" size="xs">xs</Button>
             </Panel.Body>
           </Panel>
 
           <Panel>
             <Panel.Header title="形状" />
             <Panel.Body>
-              <Button bordered shape="radius">圆角按钮</Button>
-              <Button bordered shape="round">椭圆角按钮</Button>
-              <Button bordered shape="circle">GO</Button>
-              <Button bordered shape="circle" icon={<Icon type="right" />} />
+              <Button shape="radius" theme="primary">圆角按钮</Button>
+              <Button shape="round" theme="primary">椭圆角按钮</Button>
+              <Button shape="circle" theme="primary">GO</Button>
+              <Button shape="circle" icon={<Icon type="right" />} />
             </Panel.Body>
           </Panel>